[FIX] account: recompute is_matched
In the goal of fixing the batch payments with EPD in the bank rec widget, we discovered another bug that was preventing the batch payments to change their state to "reconciled".
This was due to some compute not being triggered again when one of its dependences did change (during another compute).

task-4681366

closes odoo/odoo#222303

Related: odoo/enterprise#83077
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>