[FIX] pos_{online_payment_}self_order: enable online payments in kiosk
Currently, you are able to see online payment methods in the kiosk if
you have at least a terminal method registered. Since you're able to see
them you can also select it. Once paid you are redirected to the confirmation
screen saying that the order is being prepared but the order is not sent
to the preparation display, although the order is paid in the backend.

Steps to reproduce:
-------------------
* Create a payent method using terminal, for easier setup use stripe,
you'll only need to setup serial number on the payment method but you
do not need to do the whole stripe setup
* Create an online payment method (use demo for example)
* Change the kiosk settings to use those two payment methods
* Change the settings of the preparation display to use the kiosk
* Open preparation display
* Open kiosk, make on order, and select the online payment method
* Scan QR code and pay
> You are redirected to the confirmation page saying that the order
is being prepared while not sent to the preparation display

Why the fix:
------------
In this workflow, the order is being sent to the backend thanks to
an override in `_render_pay_confirmation` which calls `_process_order`.
However, when an exit route exists, we redirect the requests and we don't
call `_render_pay_confirmation`. We now add a method that will be overriden
in `pos_online_payment_self_order_preparation_display` and that is called
before the redirection, that will be responsible for sending the order to
the preparation display.

We also decide to only allow to use the online payment methods that
have been set on the config. Previously if two were existing but only
one setup you were able to select any of them.

We also make a change to the `filterPaymentMethods` method. This method
is called when pressing the pay button and is indirectly responsible for
the fact that if there's only one payment method then we directly start
the payment process.

Prior to this fix if we only had a terminal method setup and clicked
to pay an order we directly had the screen telling us to follow the
instructions on the terminal. However if we only had a online payment
method, when clicking on pay, we would directly have the screen prompting
us to pay at the register, we could never select the payment method.

With the fix, when there's only one payment method and it's online,
clicking on pay will show the qr code.

opw-5001998

closes odoo/odoo#223689

Related: odoo/enterprise#92825
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>