[IMP] l10n_sa_edi: consider 409 as successful submission
ZATCA introduced a new response code (409) to handle duplicate invoice submissions for B2C.
This is helpful because when the submission timesout, we are left unsure whether ZATCA
successfuly received the invoice or not. The next time Odoo tries to send the same invoice,
Zatca will respond with a 409 error if it was received earlier. In which case, we mark
the invoice as successfully sent.

A similar flow applies for B2B, with a response code of 208 for duplicate invoices.
However, for B2B, ZATCA accepts the duplicate with a warning instead of an error,
So the invoice gets marked as sent in Odoo, and no changes need to be done there.

closes odoo/odoo#224478

Task-id: 5027612
X-original-commit: 7c3d34f5ef4d111fbbc89a07e207fe5add4d0275
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Khaled Fahmy (khfa) <khfa@odoo.com>